AthletIA/requirements.txt
django==5.0.1
mssql-django==1.4
pillow==11.3.0
djangorestframework==3.16.1
requests==2.32.5
django-crispy-forms==2.4
ollama==0.6.0
django-widget-tweaks==1.5.0
joblib==1.5.2
pandas==2.3.3
numpy==2.3.4
scikit-learn==1.6.1
tensorflow==2.20.0
dill==0.4.0
google-api-python-clientS
google-auth-httplib2
google-auth-oauthlib